requirements.txt
# Core ML libraries
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.1
matplotlib==3.7.2
seaborn==0.12.2

# API and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
httpx==0.25.0

# MLOps and experiment tracking
mlflow==2.7.1
dvc==3.30.3
prometheus-client==0.18.0
structlog==23.1.0

# Data validation and testing
pydantic==2.4.2
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Development and utilities
jupyter==1.0.0
ipykernel==6.25.2
tqdm==4.66.1
black==23.11.0
isort==5.12.0
dvc==3.46.0
kagglehub==0.2.5
python-multipart==0.0.6
httpx==0.25.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
opencv-python==4.8.1.78
